requirements.txt
python-telegram-bot>=20.0
asyncpraw
aiohttp
python-dotenv
anthropic>=0.18.0